fix: makes progress tab a seprate app
