+++ openedx/features/wikimedia_features/progress_tab/settings/common.py
@@ -1,0 +1,8 @@
+
+"""Settings"""
+
+
+def plugin_settings(settings):
+    """
+    Required Common settings
+    """
