+++ openedx/features/wikimedia_features/progress_tab/tab.py
@@ -1,4 +1,3 @@
-from django.conf import settings
 from django.utils.translation import ugettext as _
 from django.utils.translation import ugettext_noop
 
@@ -10,11 +9,13 @@ class WikimediaProgressTab(TabFragmentViewMixin, EnrolledTab):
     type = 'wikimedia_progress_tab'
     title = ugettext_noop('Progress')
     priority = None
-    fragment_view_name = 'openedx.features.wikimedia_features.wikimedia_general.views.WikimediaProgressFragmentView'
+    fragment_view_name = 'openedx.features.wikimedia_features.progress_tab.views.WikimediaProgressFragmentView'
     is_hideable = False
     is_default = True
     body_class = 'wikimedia_progress'
 
     @classmethod
     def is_enabled(cls, course, user=None):
+        if not super().is_enabled(course, user=user):
+            return False
         return True
