+++ openedx/features/wikimedia_features/progress_tab/templates/progress_tab/wikimedia_progress_fragment.html
